backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
PyJWT==2.*
bcrypt==4.*
pydantic==2.*
pydantic-settings==2.*
aiosqlite==0.20.*
pytest==8.*
pytest-asyncio==0.24.*
httpx==0.28.*
